Moments for the Normal Inverse Gaussian

Computes the first four moments for the normal inverse Gaussian distribution.

nigMean(alpha = 1, beta = 0, delta = 1, mu = 0)
nigVar(alpha = 1, beta = 0, delta = 1, mu = 0)
nigSkew(alpha = 1, beta = 0, delta = 1, mu = 0)
nigKurt(alpha = 1, beta = 0, delta = 1, mu = 0)

Arguments

alpha, beta, delta, mu

are numeric values where alpha is the location parameter, beta is the location parameter, delta is the first shape parameter, and mu is the second shape parameter.

Value

All values for the *nig functions are numeric vectors: d* returns the density, p* returns the distribution function, q* returns the quantile function, and r* generates random deviates. All values have attributes named "param" listing the values of the distributional parameters.

References

Scott, D. J., Wuertz, D. and Tran, T. T. (2008) Moments of the Generalized Hyperbolic Distribution. Preprint.

Author

Diethelm Wuertz.

Examples

## nigMean -
   # Median:
   nigMean(alpha = 1, beta = 0, delta = 1, mu = 0)
#> mean 
#>    0 
 
## nigVar - 
   # Inter-quartile Range:
   nigVar(alpha = 1, beta = 0, delta = 1, mu = 0)
#> var 
#>   1 
 
## nigSKEW -  
   # Robust Skewness:
   nigSkew(alpha = 1, beta = 0, delta = 1, mu = 0)
#> skew 
#>    0 
   
## nigKurt -
   # Robust Kurtosis:
   nigKurt(alpha = 1, beta = 0, delta = 1, mu = 0)
#> kurt 
#>    3