Robust Moments for the NIG

Computes the first four robust moments for the Normal Inverse Gaussian Distribution.

nigMED(alpha = 1, beta = 0, delta = 1, mu = 0)
nigIQR(alpha = 1, beta = 0, delta = 1, mu = 0)
nigSKEW(alpha = 1, beta = 0, delta = 1, mu = 0)
nigKURT(alpha = 1, beta = 0, delta = 1, mu = 0)

Arguments

alpha, beta, delta, mu

are numeric values where alpha is the location parameter, beta is the location parameter, delta is the first shape parameter, and mu is the second shape parameter.

Value

All values for the *nig functions are numeric vectors: d* returns the density, p* returns the distribution function, q* returns the quantile function, and r* generates random deviates. All values have attributes named "param" listing the values of the distributional parameters.

Author

Diethelm Wuertz.

Examples

## nigMED -
   # Median:
   nigMED(alpha = 1, beta = 0, delta = 1, mu = 0)
#>           MED 
#> -1.058612e-07 

## nigIQR -
   # Inter-quartile Range:
   nigIQR(alpha = 1, beta = 0, delta = 1, mu = 0)
#>      IQR 
#> 1.079179 

## nigSKEW -
   # Robust Skewness:
   nigSKEW(alpha = 1, beta = 0, delta = 1, mu = 0)
#>          SKEW 
#> -8.292074e-08 

## nigKURT -
   # Robust Kurtosis:
   nigKURT(alpha = 1, beta = 0, delta = 1, mu = 0)
#>     KURT 
#> 1.387172