Slider GUI for Stable Distribution

The stableSlider() function provides interactive displays of density and probabilities of stable distributions.

Usage

stableSlider(col= "steelblue", col.med = "gray30")

Arguments

col

colour for the density and distributions functions.

col.med

colour for the median.

Value

The stableSlider() function displays densities and probabilities of the skew stable distribution, for educational purposes.

Author

Diethelm Wuertz for the Rmetrics R-port

References

see those in dstable, in package stabledist.

See also

Examples

if(dev.interactive())
   stableSlider()