Basic Time Series Statistics

Computes basic financial time series statististics.

List of Functions:

basicStats	Computes an overview of basic statistical values.

Usage

basicStats(x, ci = 0.95)

Arguments

ci

confidence interval, a numeric value, by default 0.95, i.e. 95 percent.

x

an object of class "timeSeries" or any other object which can be transformed by the function as.timeSeries into an object of class timeSeries. The latter case, other then timeSeries objects, is more or less untested.

Value

basicsStats
returns a data frame with the following entries and row names: nobs, NAs, Minimum, Maximum , 1. Quartile, 3. Quartile, Mean, Median, Sum, SE Mean, LCL Mean, UCL Mean, Variance, Stdev, Skewness, Kurtosis.

Examples

## basicStats -
   # Simulated Monthly Return Data:
   tS = timeSeries(matrix(rnorm(12)), timeDate::timeCalendar())
   basicStats(tS)
#>                  TS.1
#> nobs        12.000000
#> NAs          0.000000
#> Minimum     -1.853133
#> Maximum      2.061269
#> 1. Quartile -0.616955
#> 3. Quartile  0.704696
#> Mean         0.165781
#> Median       0.492461
#> Sum          1.989370
#> SE Mean      0.335319
#> LCL Mean    -0.572252
#> UCL Mean     0.903814
#> Variance     1.349268
#> Stdev        1.161580
#> Skewness    -0.051765
#> Kurtosis    -1.177763