Correlation Tests

Tests if two series are correlated.

Usage

correlationTest(x, y, method = c("pearson", "kendall", "spearman"), 
    title = NULL, description = NULL)
   

pearsonTest(x, y, title = NULL, description = NULL) 
kendallTest(x, y, title = NULL, description = NULL)
spearmanTest(x, y, title = NULL, description = NULL)

Arguments

x, y

numeric vectors of data values.

method

a character string naming which test should be applied.

title

an optional title string, if not specified the inputs data name is deparsed.

description

optional description string, or a vector of character strings.

Value

In contrast to R's output report from S3 objects of class "htest" a different output report is produced. The classical tests presented here return an S4 object of class "fHTEST". The object contains the following slots:

@call

the function call.

@data

the data as specified by the input argument(s).

@test

a list whose elements contain the results from the statistical test. The information provided is similar to a list object of class "htest".

@title

a character string with the name of the test. This can be overwritten specifying a user defined input argument.

@description

a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The slot @test returns an object of class "list" containing (at least) the following elements:

statistic

the value(s) of the test statistic.

p.value

the p-value(s) of the test.

parameters

a numeric value or vector of parameters.

estimate

a numeric value or vector of sample estimates.

conf.int

a numeric two row vector or matrix of 95

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

Details

The function correlationTest tests for association between paired samples allowing to compute Pearson's product moment correlation coefficient, Kendall's tau, or Spearman's rho.

Note

Some of the test implementations are selected from R's ctest package.

References

Conover, W. J. (1971); Practical nonparametric statistics, New York: John Wiley & Sons.

Lehmann E.L. (1986); Testing Statistical Hypotheses, John Wiley and Sons, New York.

Author

R-core team for hypothesis tests implemented from R's package ctest.

See also

Examples

## x, y -
   x = rnorm(50)
   y = rnorm(50)
  
## correlationTest - 
   correlationTest(x, y, "pearson")
#> 
#> Title:
#>  Pearson's Correlation Test
#> 
#> Test Results:
#>   PARAMETER:
#>     Degrees of Freedom: 48
#>   SAMPLE ESTIMATES:
#>     Correlation: 2e-04
#>   STATISTIC:
#>     t: 0.0016
#>   P VALUE:
#>     Alternative Two-Sided: 0.9987 
#>     Alternative      Less: 0.5006 
#>     Alternative   Greater: 0.4994 
#>   CONFIDENCE INTERVAL:
#>     Two-Sided: -0.2781, 0.2786
#>          Less: -1, 0.2356
#>       Greater: -0.2352, 1
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#> 
   correlationTest(x, y, "kendall")
#> 
#> Title:
#>  Kendall's tau Correlation Test
#> 
#> Test Results:
#>   SAMPLE ESTIMATES:
#>     tau: 0.009
#>   STATISTIC:
#>     z: 0.092
#>     T | Exact: 618
#>   P VALUE:
#>     Alternative         Two-Sided: 0.9267 
#>     Alternative Two-Sided | Exact: 0.9337 
#>     Alternative              Less: 0.5367 
#>     Alternative      Less | Exact: 0.5398 
#>     Alternative           Greater: 0.4633 
#>     Alternative   Greater | Exact: 0.4668 
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#> 
   spearmanTest(x, y)
#> 
#> Title:
#>  Spearman's rho Correlation Test
#> 
#> Test Results:
#>   SAMPLE ESTIMATES:
#>     rho: 0.016
#>   STATISTIC:
#>     S: 20492
#>   P VALUE:
#>     Alternative Two-Sided: 0.9122 
#>     Alternative      Less: 0.5442 
#>     Alternative   Greater: 0.4561 
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#>