Tests Class Representation and Utilities

Class representation, methods and utility functions for objects of class 'fHTEST'.

The class representation and methods are:

fHTEST	Representation for an S4 object of class "fHTEST",

show	S4 print method.

Usage

# S4 method for fHTEST
show(object)

Arguments

object

[show] -
an S4 object of class "fHTEST".

Value

In contrast to R's output report from S3 objects of class "htest" a different output report is produced. The tests return an S4 object of class "fHTEST". The object contains the following slots:

@call

the function call.

@data

the data as specified by the input argument(s).

@test

a list whose elements contain the results from the statistical test. The information provided is similar to a list object of class "htest".

@title

a character string with the name of the test. This can be overwritten specifying a user defined input argument.

@description

a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The slot @test returns an object of class "list" containing the following elements:

statistic

the value(s) of the test statistic.

p.value

the p-value(s) of the test.

parameters

a numeric value or vector of parameters.

estimate

a numeric value or vector of sample estimates.

conf.int

a numeric two row vector or matrix of 95

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

Examples

## fHTEST -
   getClass("fHTEST")
#> Class "fHTEST" [package "fBasics"]
#> 
#> Slots:
#>                                                                   
#> Name:         call        data        test       title description
#> Class:        call        list        list   character   character
   getSlots("fHTEST")
#>        call        data        test       title description 
#>      "call"      "list"      "list" "character" "character"