Two Sample Kolmogorov--Smirnov Test

Tests if two series are distributional equivalent.

Usage

ks2Test(x, y, title = NULL, description = NULL)

Arguments

x, y

numeric vectors of data values.

title

an optional title string, if not specified the inputs data name is deparsed.

description

optional description string, or a vector of character strings.

Value

In contrast to R's output report from S3 objects of class "htest" a different output report is produced. The classical tests presented here return an S4 object of class "fHTEST". The object contains the following slots:

@call

the function call.

@data

the data as specified by the input argument(s).

@test

a list whose elements contain the results from the statistical test. The information provided is similar to a list object of class "htest".

@title

a character string with the name of the test. This can be overwritten specifying a user defined input argument.

@description

a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The slot @test returns an object of class "list" containing (at least) the following elements:

statistic

the value(s) of the test statistic.

p.value

the p-value(s) of the test.

parameters

a numeric value or vector of parameters.

estimate

a numeric value or vector of sample estimates.

conf.int

a numeric two row vector or matrix of 95

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

Details

The test ks2Test performs a Kolmogorov--Smirnov two sample test that the two data samples x and y come from the same distribution, not necessarily a normal distribution. That means that it is not specified what that common distribution is.

References

Conover, W. J. (1971); Practical nonparametric statistics, New York: John Wiley & Sons.

Lehmann E.L. (1986); Testing Statistical Hypotheses, John Wiley and Sons, New York.

Author

R-core team for hypothesis tests implemented from R's package ctest.

Examples

## rnorm - 
   # Generate Series:
   x = rnorm(50)
   y = rnorm(50)
  
## ks2Test - 
   ks2Test(x, y)
#> 
#> Title:
#>  Kolmogorov-Smirnov Two Sample Test
#> 
#> Test Results:
#>   STATISTIC:
#>     D | Two Sided: 0.22
#>        D^- | Less: 0.1
#>     D^+ | Greater: 0.22
#>   P VALUE:
#>     Alternative       Two-Sided: 0.1786 
#>     Alternative Exact Two-Sided: 0.1786 
#>     Alternative            Less: 0.6065 
#>     Alternative         Greater: 0.08892 
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#>