Two Sample Location Tests

Tests if two series differ in their distributional location parameter.

Usage

locationTest(x, y, method = c("t", "kw2"), 
    title = NULL, description = NULL)

Arguments

x, y

numeric vectors of data values.

method

a character string naming which test should be applied.

title

an optional title string, if not specified the inputs data name is deparsed.

description

optional description string, or a vector of character strings.

Value

In contrast to R's output report from S3 objects of class "htest" a different output report is produced. The classical tests presented here return an S4 object of class "fHTEST". The object contains the following slots:

@call

the function call.

@data

the data as specified by the input argument(s).

@test

a list whose elements contain the results from the statistical test. The information provided is similar to a list object of class "htest".

@title

a character string with the name of the test. This can be overwritten specifying a user defined input argument.

@description

a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The slot @test returns an object of class "list" containing (at least) the following elements:

statistic

the value(s) of the test statistic.

p.value

the p-value(s) of the test.

parameters

a numeric value or vector of parameters.

estimate

a numeric value or vector of sample estimates.

conf.int

a numeric two row vector or matrix of 95

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

Details

The method="t" can be used to determine if the two sample means are equal for unpaired data sets. Two variants are used, assuming equal or unequal variances.

The method="kw2" performs a Kruskal-Wallis rank sum test of the null hypothesis that the central tendencies or medians of two samples are the same. The alternative is that they differ. Note, that it is not assumed that the two samples are drawn from the same distribution. It is also worth to know that the test assumes that the variables under consideration have underlying continuous distributions.

Note

Some of the test implementations are selected from R's ctest package.

References

Conover, W. J. (1971); Practical nonparametric statistics, New York: John Wiley & Sons.

Lehmann E.L. (1986); Testing Statistical Hypotheses, John Wiley and Sons, New York.

Author

R-core team for hypothesis tests implemented from R's package ctest.

Examples

## rnorm - 
   # Generate Series:
   x = rnorm(50)
   y = rnorm(50)
  
## locationTest -
   locationTest(x, y, "t")
#> 
#> Title:
#>  t Test
#> 
#> Test Results:
#>   PARAMETER:
#>     x Observations: 50
#>     y Observations: 50
#>     mu: 0
#>   SAMPLE ESTIMATES:
#>     Mean of x: -0.1242
#>     Mean of y: 0.2947
#>     Var  of x: 0.7276
#>     Var  of y: 1.1013
#>   STATISTIC:
#>                 T: -2.19
#>     T | Equal Var: -2.19
#>   P VALUE:
#>     Alternative Two-Sided: 0.031 
#>     Alternative      Less: 0.0155 
#>     Alternative   Greater: 0.9845 
#>     Alternative Two-Sided | Equal Var: 0.0309 
#>     Alternative      Less | Equal Var: 0.01545 
#>     Alternative   Greater | Equal Var: 0.9846 
#>   CONFIDENCE INTERVAL:
#>     Two-Sided: -0.7986, -0.0391
#>          Less: -Inf, -0.1011
#>       Greater: -0.7366, Inf
#>     Two-Sided | Equal Var: -0.7984, -0.0393
#>          Less | Equal Var: -Inf, -0.1013
#>       Greater | Equal Var: -0.7364, Inf
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#> 
   locationTest(x, y, "kw2")
#> 
#> Title:
#>  Kruskal-Wallis Two Sample Test
#> 
#> Test Results:
#>   PARAMETER:
#>     x Observations: 50
#>     y Observations: 50
#>   SAMPLE ESTIMATES:
#>     Mean of x: -0.1242
#>     Mean of y: 0.2947
#>     Var  of x: 0.7276
#>     Var  of y: 1.1013
#>   STATISTIC:
#>     KW chi-squared: 3.9693
#>   P VALUE:
#>     0.04634 
#> 
#> Description:
#>  Mon Aug  8 08:09:50 2022
#>