Two Sample Scale Tests

Tests if two series differ in their distributional scale parameter.

Usage

scaleTest(x, y, method = c("ansari", "mood"), 
    title = NULL, description = NULL)

Arguments

x, y

numeric vectors of data values.

method

a character string naming which test should be applied.

title

an optional title string, if not specified the inputs data name is deparsed.

description

optional description string, or a vector of character strings.

Value

In contrast to R's output report from S3 objects of class "htest" a different output report is produced. The classical tests presented here return an S4 object of class "fHTEST". The object contains the following slots:

@call

the function call.

@data

the data as specified by the input argument(s).

@test

a list whose elements contain the results from the statistical test. The information provided is similar to a list object of class "htest".

@title

a character string with the name of the test. This can be overwritten specifying a user defined input argument.

@description

a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The slot @test returns an object of class "list" containing (at least) the following elements:

statistic

the value(s) of the test statistic.

p.value

the p-value(s) of the test.

parameters

a numeric value or vector of parameters.

estimate

a numeric value or vector of sample estimates.

conf.int

a numeric two row vector or matrix of 95

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

Details

The method="ansari" performs the Ansari--Bradley two--sample test for a difference in scale parameters. The test returns for any sizes of the series x and y the exact p value together with its asymptotic limit.

The method="mood", is another test which performs a two--sample test for a difference in scale parameters. The underlying model is that the two samples are drawn from f(x-l) and f((x-l)/s)/s, respectively, where l is a common location parameter and s is a scale parameter. The null hypothesis is s=1.

Note

Some of the test implementations are selected from R's ctest package.

References

Conover, W. J. (1971); Practical nonparametric statistics, New York: John Wiley & Sons.

Lehmann E.L. (1986); Testing Statistical Hypotheses, John Wiley and Sons, New York.

Moore, D.S. (1986); Tests of the chi-squared type, In: D'Agostino, R.B. and Stephens, M.A., eds., Goodness-of-Fit Techniques, Marcel Dekker, New York.

Author

R-core team for hypothesis tests implemented from R's package ctest.

Examples

## rnorm - 
   # Generate Series:
   x = rnorm(50)
   y = rnorm(50)
   
## scaleTest -
   scaleTest(x, y, "ansari")
#> 
#> Title:
#>  Ansari-Bradley Test for Scale
#> 
#> Test Results:
#>   STATISTIC:
#>     AB: 1212
#>   P VALUE:
#>     Alternative Two-Sided        : 0.385 
#>     Alternative Two-Sided | Exact: 0.3904 
#>     Alternative      Less        : 0.8075 
#>     Alternative      Less | Exact: 0.8086 
#>     Alternative   Greater        : 0.1925 
#>     Alternative   Greater | Exact: 0.1952 
#>   CONFIDENCE INTERVAL:
#>     Two-Sided | Asymptotic : 0.7992, 1.8253
#>     Two-Sided |      Exact : 0.7992, 1.8253
#>     Less      | Asymptotic : 0, 1.688
#>     Less      |      Exact : 0, 1.6881
#>     Greater   | Asymptotic : 0.8382, Inf
#>     Greater   |      Exact : 0.8382, Inf
#> 
#> Description:
#>  Mon Aug  8 08:09:51 2022
#> 
   scaleTest(x, y, "mood")
#> 
#> Title:
#>  Mood Two-Sample Test of Scale
#> 
#> Test Results:
#>   STATISTIC:
#>     Z: 0.9283
#>   P VALUE:
#>     Alternative Two-Sided: 0.3533 
#>     Alternative      Less: 0.8234 
#>     Alternative   Greater: 0.1766 
#> 
#> Description:
#>  Mon Aug  8 08:09:51 2022
#>