Description File Listing

Lists the content of a description file.

listDescription(package, character.only = FALSE)

Arguments

package

a literal character or character string denoting the name of the package to be listed.

character.only

a logical indicating whether 'package' can be assumed to be character strings.

Value

prints the description file.

See also

Examples

## listDescription -
   listDescription("fBasics")
#> 
#>  fBasics Description:
#> 
#>   Package:       fBasics
#>   Title:         Rmetrics - Markets and Basic Statistics
#>   Date:          2022-09-08
#>   Version:       4021.92
#>   Authors@R:     c(person("Diethelm", "Wuertz", role="aut", comment =
#>                    "original code") , person("Tobias", "Setz", role =
#>                    c("aut"), email = "tobias.setz@live.com") ,
#>                    person("Yohan", "Chalabi", role = "aut") ,
#>                    person("Martin","Maechler", role="ctb",
#>                    email="maechler@stat.math.ethz.ch", comment = c(ORCID =
#>                    "0000-0002-8685-9910")) , person("CRAN Team", role =
#>                    "ctb") , person(given = c("Georgi", "N."), family =
#>                    "Boshnakov", role = c("cre", "ctb"), email =
#>                    "georgi.boshnakov@manchester.ac.uk") )
#>   Description:   Provides a collection of functions to explore and to
#>                    investigate basic properties of financial returns and
#>                    related quantities.  The covered fields include
#>                    techniques of explorative data analysis and the
#>                    investigation of distributional properties, including
#>                    parameter estimation and hypothesis testing. Even more
#>                    there are several utility functions for data handling
#>                    and management.
#>   Depends:       R (>= 2.15.1)
#>   Imports:       timeDate, timeSeries, stats, grDevices, graphics,
#>                    methods, utils, MASS, spatial, gss, stabledist
#>   Suggests:      interp, RUnit, tcltk
#>   LazyData:      yes
#>   License:       GPL (>= 2)
#>   Encoding:      UTF-8
#>   URL:           https://r-forge.r-project.org/scm/viewvc.php/pkg/fBasics/?root=rmetrics
#>                    )devel), https://www.rmetrics.org
#>   BugReports:    https://r-forge.r-project.org/projects/rmetrics
#>   Author:        Diethelm Wuertz [aut] (original code), Tobias Setz
#>                    [aut], Yohan Chalabi [aut], Martin Maechler [ctb]
#>                    (<https://orcid.org/0000-0002-8685-9910>), CRAN Team
#>                    [ctb], Georgi N. Boshnakov [cre, ctb]
#>   Maintainer:    Georgi N. Boshnakov <georgi.boshnakov@manchester.ac.uk>
#>   Built:         R 4.0.3; x86_64-pc-linux-gnu; 2022-10-14 14:19:28 UTC;
#>                    unix